requirements.txt
absl-py==0.9.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
afinn==0.1
astor==0.8.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
bert-for-tf2==0.13.2; python_version >= "3.5"
bert==2.2.0
cachetools==4.0.0; python_version >= "3.5" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
certifi==2019.11.28; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
chardet==3.0.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cycler==0.10.0; python_version >= "3.6"
erlastic==2.0.0
gast==0.2.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
google-auth-oauthlib==0.4.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
google-auth==1.11.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
google-pasta==0.2.0
grpcio==1.27.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
h5py==2.10.0
idna==2.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
joblib==0.14.1; python_version >= "3.5"
keras-applications==1.0.8
keras-preprocessing==1.1.0
kiwisolver==1.1.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
markdown==3.2.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.2.0" and python_version >= "3.5"
matplotlib==3.2.1; python_version >= "3.6"
nltk==3.4.5
numpy==1.18.2; python_version >= "3.5"
oauthlib==3.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
opt-einsum==3.2.0; python_version >= "3.5"
pandas==1.0.2; python_full_version >= "3.6.1" and python_version >= "3.6"
params-flow==0.7.4; python_version >= "3.5"
protobuf==3.11.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
py-params==0.8.3; python_version >= "3.5"
pyasn1-modules==0.2.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pyasn1==0.4.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pyparsing==2.4.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pyreadr==0.2.6
python-dateutil==2.8.1; python_full_version >= "3.6.1" and python_version >= "3.6"
pytz==2019.3; python_full_version >= "3.6.1"
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.23.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
rsa==4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
scikit-learn==0.22.2.post1; python_version >= "3.5"
scipy==1.4.1; python_version >= "3.6"
seaborn==0.10.0; python_version >= "3.6"
sentencepiece==0.1.85
six==1.14.0; python_full_version >= "3.6.1" and python_version >= "3.6"
tensorboard==2.0.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.3
termcolor==1.1.0
textblob==0.15.3
tqdm==4.43.0; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.2.0")
urllib3==1.25.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
vadersentiment==3.3.1
werkzeug==1.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
wheel==0.34.2; python_version >= "3" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3"
wrapt==1.12.1
